Sampling method. Optimise posterior
The previous version did not consider the prior on theta.
Also the sampling was constant on theta_0. Trivial problem,
Now the dataset samples a theta from the true prior, then samples
an x with that theta as mean.

Also optimising on the log posterior (log likel + log prior) instead of
log likelihood only.